Functions access_verification and json_verification in module smartyard.utils
